03_fastapi_cache/requirements.txt
fastapi[all]
uvicorn
redis
redis_om
python-multipart
python-jose
fastapi-cache2
